new implementation of groups

+++ kanidm_operator/deploy/user.py
@@ -2,11 +2,8 @@ from logging import Logger
 
 import kopf
 from kanidm_operator.typing.user import UserResource
-from kanidm_operator.deployer import Deployer
-from kanidm_operator.deployer import slugify
 
 from .util import KanidmCLIClient
-import subprocess
 
 @kopf.on.create("kanidm.github.io", "v1alpha1", "users")
 @kopf.on.update("kanidm.github.io", "v1alpha1", "users")
@@ -23,6 +20,10 @@ async def on_create_user(
     cli_client.create_user(spec['name'], spec['displayName'])
     cli_client.set_user_emails(spec['name'], spec['emails'])
 
+@kopf.on.field("kanidm.github.io", "v1alpha1", "users", field="spec.name")
+async def on_update_group_name(**kwargs):
+    raise kopf.PermanentError("User name cannot be changed")
+
 @kopf.on.delete("kanidm.github.io", "v1alpha1", "users")
 async def on_delete_user(
     spec: UserResource,
@@ -33,5 +34,7 @@ async def on_delete_user(
     **kwargs,
 ):
     logger.info(f"Trying to add user {spec['name']} to kanidm in the namespace {namespace}")
-    cli_client = KanidmCLIClient(spec["kanidmName"], namespace, logger)
-    cli_client.delete_user(spec['name'])
+    # If kanidm is already gone, then don't worry about deleting
+    cli_client = KanidmCLIClient(spec["kanidmName"], namespace, logger, silence_missing_kanidm=True)
+    if cli_client.kanidm_spec is not None:
+        cli_client.delete_user(spec['name'])
